refactor: organize loaders into subfolder and rename dataset builders
