+++ simba/workflows/inference.py
@@ -16,7 +16,7 @@ import simba.core.data.molecule_pairs
 import simba.core.data.spectrum
 from simba.core.chemistry.mces_loader.load_mces import LoadMCES
 from simba.core.data.molecule_pairs import MoleculePairsOpt
-from simba.core.data.multitask_loader import MultitaskDataLoader
+from simba.core.data.multitask_dataset_builder import MultitaskDataLoader
 from simba.core.data.preprocessing_simba import PreprocessingSimba
 from simba.core.models.embedder_multitask import EmbedderMultitask
 from simba.core.training.train_utils import TrainUtils
