+++ simba/workflows/training.py
@@ -18,7 +18,7 @@ import simba.core.data.molecule_pairs
 import simba.core.data.spectrum
 from simba.core.chemistry.mces_loader.load_mces import LoadMCES
 from simba.core.data.molecule_pairs import MoleculePairsOpt
-from simba.core.data.multitask_loader import MultitaskDataLoader
+from simba.core.data.multitask_dataset_builder import MultitaskDataLoader
 from simba.core.data.preprocessing_simba import PreprocessingSimba
 from simba.core.data.sampling.custom_weighted_random_sampler import (
     CustomWeightedRandomSampler,
